Support entire Aluminum API with MPI backend.
Reorganize some code.
