+++ include/aluminum/mpi/CMakeLists.txt
@@ -1,0 +1,22 @@
+set_source_path(THIS_DIR_HEADERS
+  allgather.hpp
+  allreduce.hpp
+  alltoall.hpp
+  bcast.hpp
+  communicator.hpp
+  gather.hpp
+  reduce.hpp
+  reduce_scatter.hpp
+  scatter.hpp
+  pt2pt.hpp
+  utils.hpp
+  )
+
+set(THIS_DIR_HEADERS_TO_INSTALL ${THIS_DIR_HEADERS})
+
+# Custom install
+install(FILES ${THIS_DIR_HEADERS_TO_INSTALL}
+  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/aluminum/mpi)
+
+# Propagate the files up the tree
+set(HEADERS ${HEADERS} ${THIS_DIR_HEADERS} PARENT_SCOPE)
