+++ test/test_alltoall.cpp
@@ -148,8 +148,7 @@ int main(int argc, char** argv) {
   parse_args(argc, argv, backend, start_size, max_size);
 
   if (backend == "MPI") {
-    std::cerr << "Alltoall not supported on MPI backend." << std::endl;
-    std::abort();
+    test_correctness<Al::MPIBackend>();
 #ifdef AL_HAS_NCCL
   } else if (backend == "NCCL") {
     std::cerr << "Alltoall not supported on NCCL backend." << std::endl;
