+++ test/test_bcast.cpp
@@ -126,8 +126,7 @@ int main(int argc, char** argv) {
   parse_args(argc, argv, backend, start_size, max_size);
 
   if (backend == "MPI") {
-    std::cerr << "Bcast not supported on MPI backend." << std::endl;
-    std::abort();
+    test_correctness<Al::MPIBackend>();
 #ifdef AL_HAS_NCCL
   } else if (backend == "NCCL") {
     test_correctness<Al::NCCLBackend>();
