+++ test/test_reduce.cpp
@@ -149,8 +149,7 @@ int main(int argc, char** argv) {
   parse_args(argc, argv, backend, start_size, max_size);
 
   if (backend == "MPI") {
-    std::cerr << "Reduce not supported on MPI backend." << std::endl;
-    std::abort();
+    test_correctness<Al::MPIBackend>();
 #ifdef AL_HAS_NCCL
   } else if (backend == "NCCL") {
     test_correctness<Al::NCCLBackend>();
